Ban quote_with_schema on MySQL
