+++ migration-engine/connectors/sql-migration-connector/src/sql_renderer.rs
@@ -1,10 +1,12 @@
+#![deny(missing_docs)]
+
 mod common;
 mod mssql_renderer;
 mod mysql_renderer;
 mod postgres_renderer;
 mod sqlite_renderer;
 
-pub(crate) use common::{IteratorJoin, Quoted, QuotedWithSchema};
+pub(crate) use common::IteratorJoin;
 
 use crate::{
     database_info::DatabaseInfo,
@@ -13,41 +15,17 @@ use crate::{
     },
     sql_schema_differ::SqlSchemaDiffer,
 };
-use sql_schema_describer::walkers::{ColumnWalker, TableWalker};
-use sql_schema_describer::*;
-use std::{borrow::Cow, fmt::Write as _};
+use common::{Quoted, QuotedWithSchema};
+use sql_schema_describer::{
+    walkers::{ColumnWalker, TableWalker},
+    ColumnTypeFamily, DefaultValue, ForeignKey, SqlSchema,
+};
+use std::borrow::Cow;
 
 pub(crate) trait SqlRenderer {
     fn quote<'a>(&self, name: &'a str) -> Quoted<&'a str>;
 
-    fn quote_with_schema<'a, 'b>(&'a self, name: &'b str) -> QuotedWithSchema<'a, &'b str>;
-
-    fn render_add_foreign_key(&self, add_foreign_key: &AddForeignKey) -> String {
-        let AddForeignKey { foreign_key, table } = add_foreign_key;
-        let mut add_constraint = String::with_capacity(120);
-
-        write!(
-            add_constraint,
-            "ALTER TABLE {table} ADD ",
-            table = self.quote_with_schema(table)
-        )
-        .unwrap();
-
-        if let Some(constraint_name) = foreign_key.constraint_name.as_ref() {
-            write!(add_constraint, "CONSTRAINT {} ", self.quote(constraint_name)).unwrap();
-        }
-
-        write!(
-            add_constraint,
-            "FOREIGN KEY ({})",
-            foreign_key.columns.iter().map(|col| self.quote(col)).join(", ")
-        )
-        .unwrap();
-
-        add_constraint.push_str(&self.render_references(&table, &foreign_key));
-
-        add_constraint
-    }
+    fn render_add_foreign_key(&self, add_foreign_key: &AddForeignKey) -> String;
 
     fn render_alter_enum(&self, alter_enum: &AlterEnum, differ: &SqlSchemaDiffer<'_>) -> anyhow::Result<Vec<String>>;
 
@@ -93,5 +71,6 @@ pub(crate) trait SqlRenderer {
     /// Render a `RedefineTables` step.
     fn render_redefine_tables(&self, tables: &[String], differ: SqlSchemaDiffer<'_>) -> Vec<String>;
 
+    /// Render a table renaming step.
     fn render_rename_table(&self, name: &str, new_name: &str) -> String;
 }
